Update readme-vars.yml
Added `xpack.security.enabled=true` in ElasticSearch environment to enable the option to add/change the default ElasticSearch user passwords and prevent the `It doesn't look like the X-Pack security feature is enabled on this Elasticsearch node. Please check if you have enabled X-Pack security in your elasticsearch` error.

Added instructions on how to edit `Constants.php` to allow diskover-web to communicate with ElasticSearch. The env vars `ES_HOST` and `ES_PORT` in the docker-compose are not modifying `Constants.php` so without manually modifying this file a user will get an error `cannot connect to localhost:9200`

Clarified the proper command a user will need to run to manually generate an index.